place template linker back in files module
